+++ packages/storefront/components/s-product-card.html
@@ -12,7 +12,7 @@
         <div class="font-bold">
             {#product.price#}
         </div>
-        <s-btn-primary class="w-full" label="Add to cart" @click="action.addToCart(product.id)" />
+        <s-btn class="w-full" label="Add to cart" @click="action.addToCart(product.id)" />
         <span>{#state?.cart#}</span>
     </div>
 </div>
